// src/app/Types/Kaspi/ParamsForKaspiOrderDeliveryPointOfServiceAddressesType.ts
export default interface ParamsForKaspiOrderDeliveryPointOfServiceAddressesType {
    streetName: string | null;
    streetNumber: string | null;
    town: string | null;
    district: string | null;
    building: string | null;
    formattedAddress: string | null;
    kaspi_order_delivery_point_of_service_id: number;
    created_at: Date;
    updated_at: Date;
}